requirements.txt
flask>=3.0.0
google-play-scraper>=1.2.7
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
